+++ applications/example/server/api/auth/register.json/post.js
@@ -1,7 +1,5 @@
 ﻿module.exports = function(client, callback) {
-  application.security.register(client, client.fields.Email, client.fields.Password, function(err, user) {
-    if (user) client.context.data = { Result: 'Ok' };
-    else client.context.data = { Result: 'Error' };
-    callback();
+  application.security.register(client, client.fields.email, client.fields.password, function(err, user) {
+    callback({ result: user ? 'ok' : 'error' });
   });
 };