+++ applications/example/server/api/auth/signIn.json/post.js
@@ -1,8 +1,6 @@
 ﻿module.exports = function(client, callback) {
   application.security.signIn(client, client.fields.Login, client.fields.Password, function(isSuccess) {
-    if (isSuccess) client.context.data = { Result: 'Ok' };
-    else client.context.data = { Result: 'Error' };
     if (client.fields.loginForm) client.redirect('/');
-    callback();
+    callback({ result: isSuccess ? 'ok' : 'error' });
   });
 };