+++ applications/example/server/api/auth/signOut.json/post.js
@@ -1,4 +1,4 @@
 ﻿module.exports = function(client, callback) {
-  client.context.data = { Result: 'Ok' };
+  client.context.data = { result: 'ok' };
   application.security.signOut(client, callback);
 };