+++ applications/example/server/examples/events/sendEvent.json/get.js
@@ -1,4 +1,4 @@
 module.exports = function(client, callback) {
   application.frontend.emit('test', { data: 'data' });
-  callback('Ok');
+  callback('ok');
 };