+++ applications/example/server/examples/pgsql/getLanguages.json/get.js
@@ -1,5 +1,5 @@
 module.exports = function(client, callback) {
   aliasNamePg.query('select * from SystemLanguage where LanguageId<$1', ['10'], function(err, result) {
-    callback({ rows:result.rows, fields:result.fields });
+    callback({ rows: result.rows, fields: result.fields });
   });
 };