+++ applications/example/server/examples/simple/jsonpGet.jsonp/get.js
@@ -1,7 +1,7 @@
 ﻿module.exports = function(client, callback) {
   callback({
-    query:  client.query,
-    path:   client.path,
+    query: client.query,
+    path: client.path,
     fields: client.fields
   });
 };