+++ deploy/centos6x64.sh
@@ -4,15 +4,15 @@ yum -y install mc
 yum -y install wget
 yum -y groupinstall "Development Tools"
 cd /usr/src
-wget http://nodejs.org/dist/v4.0.0/node-v4.0.0.tar.gz
-tar zxf node-v4.0.0.tar.gz
-rm -f ./node-v4.0.0.tar.gz
-cd node-v4.0.0
+wget http://nodejs.org/dist/v4.1.0/node-v4.1.0.tar.gz
+tar zxf node-v4.1.0.tar.gz
+rm -f ./node-v4.1.0.tar.gz
+cd node-v4.1.0
 ./configure
 make
 make install
 cd ~
-rm -rf /usr/src/node-v4.0.0
+rm -rf /usr/src/node-v4.1.0
 ln -s /usr/local/bin/node /bin
 ln -s /usr/local/bin/npm /bin
 cat >/etc/yum.repos.d/mongodb.repo <<EOL